shared/requirements.txt
fastapi[standard]==0.124.0
uvicorn[standard]==0.38.0
pydantic-settings==2.12.0
structlog==25.5.0
paramiko==4.0.0
tenacity==9.1.2
sqlalchemy==2.0.44
asyncpg==0.31.0
aioboto3==15.5.0 # used for s3-like buckets
packaging==25.0
requests==2.32.5 # used internally by google-auth but not automatically installed
google-auth==2.43.0
py-cpuinfo==9.0.0
pillow==12.0.0